requirements.txt
bs4
PyYAML
jinja2
requests
tabulate
